java-connectors/kafka-connect-common/src/main/java/io/lenses/streamreactor/common/util/ByteConverters.java
/*
 * Copyright 2017-2024 Lenses.io Ltd
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package io.lenses.streamreactor.common.util;

import java.io.ByteArrayOutputStream;
import java.io.IOException;
import java.io.ObjectOutputStream;

/**
 * Class with converters to/from bytes.
 */
public class ByteConverters {

  /**
   * Converts Object to byte array.
   *
   * @param obj {@link Object} to convert.
   * @return array of bytes that represents the object or IOException if operation failed
   */
  public static byte[] toBytes(Object obj) throws IOException {
    ByteArrayOutputStream boas = new ByteArrayOutputStream();

    try (ObjectOutputStream ois = new ObjectOutputStream(boas)) {
      ois.writeObject(obj);
      return boas.toByteArray();
    }
  }
}
